requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
astroid==2.15.8
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bandit==1.7.9
black==24.4.2
blinker==1.8.2
boltons==21.0.0
bracex==2.4
build==1.2.1
CacheControl==0.14.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-option-group==0.5.6
cloudinary==1.40.0
colorama==0.4.6
constantly==23.10.4
coverage==7.5.4
crashtest==0.4.1
cryptography==42.0.8
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-stubs==5.0.2
django-stubs-ext==5.0.2
django-user-agents==0.4.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.0
dodgy==0.2.1
dulwich==0.21.7
exceptiongroup==1.2.1
executing==2.0.1
face==20.1.1
fastjsonschema==2.20.0
filelock==3.15.4
flake8==7.1.0
flake8-polyfill==1.0.2
Flask==3.0.3
frozenlist==1.4.1
geoip2==4.8.0
gitdb==4.0.11
GitPython==3.1.43
glom==22.1.0
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.0.0
incremental==22.10.0
iniconfig==2.0.0
installer==0.7.0
ipython==8.26.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
lazy-object-proxy==1.10.0
lockfile==0.12.2
mailjet-rest==1.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
maxminddb==2.6.2
mccabe
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
multidict==6.0.5
mypy==1.10.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pbr==6.0.0
peewee==3.17.6
pep8-naming==0.14.1
pexpect==4.9.0
phonenumbers==8.13.40
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.47
prospector==0.12.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==1.7.1
pylint==2.17.4
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==2.5.5
pylint-flask==0.6
pylint-plugin-utils==0.8.2
pyOpenSSL==24.1.0
pyotp==2.9.0
pyproject_hooks==1.1.0
pyrsistent==0.20.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-mock==3.14.0
python-lsp-jsonrpc==1.1.2
python3-openid==3.2.0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.9.4
redis==5.0.7
referencing==0.35.1
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-detector==0.7
rich==13.7.1
rpds-py==0.19.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
semgrep==1.79.0
semver==3.0.2
service-identity==24.1.0
setoptconf==0.3.0
setoptconf-tmp==0.3.1
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.5.0
stack-data==0.6.3
stevedore==5.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.0
traitlets==5.14.3
trove-classifiers==2024.7.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-docutils==0.21.0.20240710
types-Markdown==3.6.0.20240316
types-peewee==3.17.5.20240626
types-psycopg2==2.9.21.20240417
types-Pygments==2.18.0.20240506
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240622
types-setuptools==70.3.0.20240710
types-urllib3==1.26.25.14
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
ujson==5.10.0
urllib3==2.2.2
user-agents==2.2.0
uvicorn==0.30.1
virtualenv==20.26.3
wcmatch==8.5.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
whitenoise==6.7.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
zope.interface==6.4.post2
